refactor(material): extract InspectPanel.svelte from MaterialPopup
Move the isotopic-enrichment quick-access row (rendered when the search
query parses to known elements) into its own component.

Props: query, currentEnrichment, onenrichment. Derivation of queryElements
moves with it. No behavior change — parent renders InspectPanel
unconditionally; the component renders nothing when queryElements is empty.

Refs: #64 Phase 0
